+++ src/core/config.ts
@@ -44,7 +44,7 @@ const projectRoot = path.resolve(__dirname, "..", "..");
 
 export const config = {
   botName: "Syndicate Bot",
-  version: "1.1.1",
+  version: "1.1.2",
   author: "Ranajoy Roy",
 
   token: requireEnv("DISCORD_TOKEN"),